backend_python/requirements.txt
flask
flask-cors
python-docx
PyMuPDF
Babel